Struct ucl::object::builder::Builder

pub struct Builder {
    // some fields omitted
}

Build element object.

This structure is immutable typed reference to object inside parsed tree. It can be one of Type elements and can be cast only to given type.

Methods

impl Builder

fn from_ptr(obj: *mut ucl_object_t) -> Option<Self>

Create new Object form raw pointer. Internal use only.

fn build(self) -> Object